# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RiPto Sells API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current auth user
 */
export const GetCurrentAuthUserResponse = zod.object({
  user: zod
    .object({
      id: zod.string(),
      username: zod.string().optional(),
      firstName: zod.string().optional(),
      lastName: zod.string().optional(),
      profileImage: zod.string().optional(),
    })
    .nullable(),
});

/**
 * @summary Get all categories
 */
export const GetCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string().optional(),
  imageUrl: zod.string().optional(),
  productCount: zod.number().optional(),
});
export const GetCategoriesResponse = zod.array(GetCategoriesResponseItem);

/**
 * @summary Get all products
 */
export const GetProductsQueryParams = zod.object({
  categoryId: zod.coerce.number().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.boolean().optional(),
});

export const GetProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  originalPrice: zod.number().optional(),
  imageUrl: zod.string().optional(),
  images: zod.array(zod.string()).optional(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  rating: zod.number().optional(),
  reviewCount: zod.number().optional(),
  inStock: zod.boolean().optional(),
  featured: zod.boolean().optional(),
  badge: zod.string().optional(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  originalPrice: zod.number().optional(),
  imageUrl: zod.string().optional(),
  images: zod.array(zod.string()).optional(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  rating: zod.number().optional(),
  reviewCount: zod.number().optional(),
  inStock: zod.boolean().optional(),
  featured: zod.boolean().optional(),
  badge: zod.string().optional(),
});

/**
 * @summary Get cart items
 */
export const GetCartResponseItem = zod.object({
  id: zod.number(),
  productId: zod.number(),
  quantity: zod.number(),
  product: zod
    .object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string().optional(),
      price: zod.number(),
      originalPrice: zod.number().optional(),
      imageUrl: zod.string().optional(),
      images: zod.array(zod.string()).optional(),
      categoryId: zod.number(),
      categoryName: zod.string().optional(),
      rating: zod.number().optional(),
      reviewCount: zod.number().optional(),
      inStock: zod.boolean().optional(),
      featured: zod.boolean().optional(),
      badge: zod.string().optional(),
    })
    .optional(),
});
export const GetCartResponse = zod.array(GetCartResponseItem);

/**
 * @summary Add item to cart
 */
export const AddToCartBody = zod.object({
  productId: zod.number(),
  quantity: zod.number(),
});

/**
 * @summary Update cart item quantity
 */
export const UpdateCartItemParams = zod.object({
  productId: zod.coerce.number(),
});

export const UpdateCartItemBody = zod.object({
  quantity: zod.number(),
});

export const UpdateCartItemResponse = zod.object({
  id: zod.number(),
  productId: zod.number(),
  quantity: zod.number(),
  product: zod
    .object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string().optional(),
      price: zod.number(),
      originalPrice: zod.number().optional(),
      imageUrl: zod.string().optional(),
      images: zod.array(zod.string()).optional(),
      categoryId: zod.number(),
      categoryName: zod.string().optional(),
      rating: zod.number().optional(),
      reviewCount: zod.number().optional(),
      inStock: zod.boolean().optional(),
      featured: zod.boolean().optional(),
      badge: zod.string().optional(),
    })
    .optional(),
});

/**
 * @summary Remove item from cart
 */
export const RemoveFromCartParams = zod.object({
  productId: zod.coerce.number(),
});

export const RemoveFromCartResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Clear all cart items
 */
export const ClearCartResponse = zod.object({
  success: zod.boolean(),
});