# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RiPto Sells API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product?: Product;
}